backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
faiss-cpu==1.7.4
